Add autocomplete functionality to search
Two new public and private methods were added to the AutocompleteHelper class, in order to support the creation and parsing of autocomplete items for JavaScript usage. The beta search view was also updated to utilize the new autocomplete functionality. In addition, the target framework of the OCRQueueManager project file has been changed from 'net8.0-windows' to 'net8.0'.